requirements.txt
psycopg2-binary
faker
pandas
python-dotenv
